Refactor (#11)
* rearrange imports and relocate functions

* factor out linux automation

* remove redundant imports

* Add comments and factor out OS + mount check

* add comment

* fix comment

* fix docstring typo

* add newlines between arg checks
