+++ kobo_backup.py
@@ -1,195 +1,52 @@
-import json
-import os
-import signal
+import argparse
 import datetime
+import glob
+import os
 from pathlib import Path
 import platform
-import glob
+import shutil
 import subprocess
 import sys
-import shutil
-import platform
-import argparse
 
-
-def get_directory_size(directory):  # figure out how much was backed up.
-    total = 0
-    try:
-        for entry in os.scandir(directory):
-            if entry.is_file():
-                total += entry.stat().st_size
-            elif entry.is_dir():
-                total += get_directory_size(entry.path)
-    except NotADirectoryError:
-        return os.path.getsize(directory)
-    except PermissionError:
-        return 0
-    return total
-
-
-def get_size_format(
-    b, factor=1024, suffix="B"
-):  # convert bytes to something human readable.
-    for unit in ["", "K", "M", "G", "T", "P", "E", "Z"]:
-        if b < factor:
-            return f"{b:.2f}{unit}{suffix}"
-        b /= factor
-    return f"{b:.2f}Y{suffix}"
-
+from automation.automation_utils import automate_for_linux
+from utils import get_directory_size, get_size_format, get_user_os_and_kobo_mountpoint
 
 def main(args):
-    label = "KOBOeReader"  # volume label of kobo - this is the default across models but could change in the future.
+
+    # Volume label of kobo - this is the default across models but could change in the future.
+    label = "KOBOeReader"
+    # the folder in which backups will be placed. This is OS agnostic.
     backup_base_directory = str(
         os.path.join(os.path.expanduser("~"), "Backups", "kobo")
-    )  # the folder in which backups will be placed. This should be OS agnostic.
+    )  
 
-    # Check status of auto backup
-    if args.status:
+    # Check if user is trying to set up automation.
+    # This check could just be `if len(sys.argv) > 1` but doing it like this makes it easier to add additional arguments for other features later.
+    if args.auto or args.cancel or args.disable or args.enable or args.status:
         if platform.system() == "Linux":
-            # Check if auto backup is enabled
-            try:
-                # Check if watcher script is running (will throw subprocess.CalledProcessError if not)
-                subprocess.check_output(["pgrep", "-f", "linux_automation"])
-                print("Auto backup is currently enabled (watching for Kobo...).")
-            except subprocess.CalledProcessError:
-                print("Auto backup is currently disabled (watcher script not detected).")
-            # Check of the autostart file exists
-            autostart_path = os.path.expanduser("~/.config/autostart/")
-            desktop_file_name = "auto_kobo_backup.desktop"
-            if os.path.exists(autostart_path + desktop_file_name):
-                print("Auto backup is enabled on restart.")
-            else:
-                print("Auto backup script will not run on restart.")
-            sys.exit()
-        else:
-            print(
-                "The automation feature is currently only supported on Linux. Exiting...."
-            )
-            sys.exit()
-    # Setup auto backup
-    if args.auto:
-        if platform.system() == "Linux":
-            from utils import create_linux_autostart_script
-            # Create the autostart script
-            create_linux_autostart_script()
-            sys.exit()
-        else:
-            print(
-                "The automation feature is currently only supported on Linux. Exiting...."
-            )
-            sys.exit()
-    # Temporarily disable auto_backup
-    if args.disable:
-        if platform.system() == "Linux":
-            try:
-                # Find pid of running watcher script
-                pid = (
-                    subprocess.check_output(["pgrep", "-f", "linux_automation"])
-                    .decode("utf-8")
-                    .strip()
-                )
-                # Kill it
-                os.kill(int(pid), signal.SIGTERM)
-                sys.exit()
-            except subprocess.CalledProcessError: # Command '['pgrep', '-f', 'linux_automation']' will return non-zero exit status 1 if no process is found.
-                print("No auto backup is currently running.")
-                sys.exit()
-        else:
-            print(
-                "The automation feature is currently only supported on Linux. Exiting...."
-            )
-            sys.exit()
-    # Re-enable auto backup (Or enable just for this session)
-    if args.enable:
-        if platform.system() == "Linux":
-            # Run the automation script
-            path_to_automation_script = os.getcwd() + os.sep + "linux_automation.py"
-            subprocess.run(["chmod", "+x", path_to_automation_script])
-            print(
-                "Running automation script temporarily in this terminal session... Connect your kobo."
-            )
-            subprocess.run([sys.executable, path_to_automation_script])
-        else:
-            print(
-                "The automation feature is currently only supported on Linux. Exiting...."
-            )
-            sys.exit()
-    # Cancel auto backup
-    if args.cancel:
-        # Remove the autostart script
-        if platform.system() == "Linux":
-            autostart_path = os.path.expanduser("~/.config/autostart/")
-            desktop_file_name = "auto_kobo_backup.desktop"
-            try:
-                os.remove(autostart_path + desktop_file_name)
-                print(
-                    "Cancelled auto-backup (removed file in autostart called "
-                    + desktop_file_name
-                    + ")"
-                )
-                try:
-                    # Find pid of running watcher script
-                    pid = (
-                        subprocess.check_output(["pgrep", "-f", "linux_automation"])
-                        .decode("utf-8")
-                        .strip()
-                    )
-                    # Kill it
-                    os.kill(int(pid), signal.SIGTERM)
-                    print("Killed running watcher script as well.")
-                except subprocess.CalledProcessError:
-                    pass
-            except FileNotFoundError:
-                print("There was no auto backup set up.")
-            sys.exit()
+            automate_for_linux(args)
         else:
             print(
                 "The automation feature is currently only supported on Linux. Exiting...."
             )
             sys.exit()
 
-    if platform.system() == "Windows":  # Get mount point on Windows
-        import wmi
+    system_info = get_user_os_and_kobo_mountpoint(label)
 
-        # Set up WMI object for later
-        c = wmi.WMI()
-        kobos = []
-        # Get all drives and their infos
-        for drive in c.Win32_LogicalDisk():
-            # If any drive is called the label, append it to the list
-            if drive.VolumeName == label:
-                kobos.append(drive.Name + os.sep)
-        user_os = "Windows"
-    elif platform.system() == "Linux":  # Get mount point on Linux
-        lsblk_check = subprocess.check_output(["lsblk", "-f", "--json"]).decode("utf8")
-        lsblk_json = json.loads(lsblk_check)
-        kobos = [
-            device
-            for device in lsblk_json["blockdevices"]
-            if device.get("label", None) == label
-        ]
-        kobos = [kobo["mountpoint"] for kobo in kobos]
-        user_os = "Linux"
-    elif platform.system() == "Darwin":  # Get mount point on MacOS
-        df_output = subprocess.check_output(("df", "-Hl")).decode("utf8")
-        output_parts = [o.split() for o in df_output.split("\n")]
-        kobos = [o[-1] for o in output_parts if f"/Volumes/{label}" in o]
-        user_os = "macOS"
-    else:
-        raise Exception(f"Unsupported OS: {platform.system()=} {platform.release()=}")
-
-    if len(kobos) > 1:
-        raise RuntimeError(f"Multiple Kobo devices detected: {kobos}.")
-    elif len(kobos) == 0:
+    # Check we have one and only one Kobo connected
+    if len(system_info.kobos) > 1:
+        raise RuntimeError(f"Multiple Kobo devices detected: {system_info.kobos}.")
+    elif len(system_info.kobos) == 0:
         print("No kobos detected.")
         sys.exit()
     else:
-        [kobo] = kobos
-        print(f"Kobo mountpoint is: {Path(kobo)} on {user_os}.")
+        [kobo] = system_info.kobos
+        print(f"Kobo mountpoint is: {Path(kobo)} on {system_info.user_os}.")
 
+    # Check backup base directory exists locally, if not create it.
     backup_folder_exists = os.path.isdir(
         backup_base_directory
-    )  # check backup base directory exists locally, if not create it.
+    ) 
     if not backup_folder_exists:
         print(f"No backup folder detected. Creating {backup_base_directory}.")
         os.makedirs(backup_base_directory)
@@ -198,28 +55,33 @@ def main(args):
             f"An existing kobo backup folder was detected at {backup_base_directory}."
         )
 
-    try:
-        previous_backup = max(
-            glob.glob(os.path.join(backup_base_directory, "*/")), key=os.path.getmtime
-        )  # get the folder of the previous backup that occured
-    except ValueError:
-        pass
-
+    # Append datestamp to directory name.
     backup_path = os.path.join(
         backup_base_directory,
         "kobo_backup_" + datetime.datetime.now().strftime("%Y-%m-%d_%H-%M"),
-    )  # append datestamp to directory name.
+    )  
+    # Check that we haven't already backed up during this minute.
     if os.path.isdir(backup_path):
         print(
             f"A backup of the kobo was already completed at {datetime.datetime.now().strftime('%Y-%m-%d %H:%M')}. Try again in a minute."
         )
         sys.exit()
 
-    try:  # copy files
+    # Get the folder of the previous backup that occured
+    try:
+        previous_backup = max(
+            glob.glob(os.path.join(backup_base_directory, "*/")), key=os.path.getmtime
+        )  
+    except ValueError:
+        pass
+
+    # Copy files
+    try: 
         shutil.copytree(Path(kobo), backup_path)
     except OSError:  # some unrequired .Trashes will return 'operation not permitted'.
         pass
 
+    # Print size of last backup and current backup to stdout.
     try:
         previous_backup
         print(
@@ -231,8 +93,9 @@ def main(args):
     print(
         f"Backup complete. Copied {sum(len(files) for _, _, files in os.walk(backup_path))} files with a size of {get_size_format(get_directory_size(backup_path))} to {backup_path}."
     )
+
+    # Only tested the below on Linux
     try:
-        # Only tested the below on Linux
         # Open a notification to say it was backed up
         subprocess.Popen(["notify-send", f"Backed up!"])
         # Open the file explorer to the backed up directory
